+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * `step_sequence_onehot()` now takes tokenlists as input.
 * added {tokenizers.bpe} engine to `step_tokenize()`.
+* added {udpipe} engine to `step_tokenize()`.
 
 # textrecipes 0.3.0